Test cases for TabbedFormTabs
